+++ pdata/internal/cmd/pdatagen/internal/metrics_structs.go
@@ -267,6 +267,7 @@ var numberDataPoint = &messageValueStruct{
 		},
 		exemplarsField,
 		dataPointFlagsField,
+		dataPointFlagsFieldStruct,
 	},
 }
 
@@ -289,6 +290,7 @@ var histogramDataPoint = &messageValueStruct{
 		explicitBoundsField,
 		exemplarsField,
 		dataPointFlagsField,
+		dataPointFlagsFieldStruct,
 		&optionalPrimitiveValue{
 			fieldName:        "Min",
 			fieldType:        "Double",
@@ -364,6 +366,7 @@ var exponentialHistogramDataPoint = &messageValueStruct{
 		},
 		exemplarsField,
 		dataPointFlagsField,
+		dataPointFlagsFieldStruct,
 		&optionalPrimitiveValue{
 			fieldName:        "Min",
 			fieldType:        "Double",
@@ -423,6 +426,7 @@ var summaryDataPoint = &messageValueStruct{
 			returnSlice:     quantileValuesSlice,
 		},
 		dataPointFlagsField,
+		dataPointFlagsFieldStruct,
 	},
 }
 
@@ -486,6 +490,15 @@ var exemplar = &messageValueStruct{
 		},
 		traceIDField,
 		spanIDField,
+	},
+}
+
+var dataPointFlagsFieldStruct = &messageValueField{
+	fieldName:       "FlagsStruct",
+	originFieldName: "Flags",
+	returnMessage: &messageValueStruct{
+		structName:     "MetricDataPointFlagsStruct",
+		originFullName: "uint32",
 	},
 }